ml-pipeline/requirements.txt
# Core ML
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
scikit-learn==1.4.0
xgboost==2.0.3
torch-geometric==2.5.0

# GCP
google-cloud-aiplatform==1.42.1
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.0
google-auth==2.27.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==15.0.0
scipy==1.12.0

# Geospatial
shapely==2.0.2
geopy==2.4.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Monitoring
tensorboard==2.15.1
wandb==0.16.2

